+++ ReleaseNotes.md
@@ -1,6 +1,9 @@
 * **\[MDAPI-216]\[C++][Console]** Added PriceLevelBook sample
 * **\[MDAPI-76]\[C++]** Implemented TextMessage event
 * Migrated to Graal SDK v2.2.1
+* [MDAPI-214][C++][Console] Added MarketDepthModelSample
+    * Added `MarketDepthModelSample`
+    * Added `CmdArgsUtils::parseEventSources` method.
 
 ## v4.0.0